+++ src/wasm-s-parser.h
@@ -212,6 +212,7 @@ private:
   Expression* makeSIMDTernary(Element& s, SIMDTernaryOp op);
   Expression* makeSIMDShift(Element& s, SIMDShiftOp op);
   Expression* makeSIMDLoad(Element& s, SIMDLoadOp op);
+  Expression* makeSIMDLoadStoreLane(Element& s, SIMDLoadStoreLaneOp op);
   Expression* makeMemoryInit(Element& s);
   Expression* makeDataDrop(Element& s);
   Expression* makeMemoryCopy(Element& s);